+++ src/app/dice/dice.component.ts
@@ -50,6 +50,7 @@ export class DiceComponent implements OnInit {
     roll(amount: number, size: number) {
         this.diceService.roll(amount, size, this.bonus, this.characterService, false);
         this.bonus = 0;
+        this.characterService.process_ToChange();
     }
 
     get_Creature(creatureType: string) {