+++ Changelog.md
@@ -6,3 +6,8 @@
 ## 1.0.1
 
 - After applying some errata, certain containers have been removed, and all characters now receive a second default inventory for worn tools.
+
+## 1.0.2
+
+- Monks created before version 1.0.2 will lose their skill increases from Path to Perfection, and gain matching feature choices instead.
+- A Foundry VTT integration allows you to send your dice rolls to a FVTT session. This requires the External Dice Roll Connector module to be enabled in FVTT.